Make engine base classes public, move `_merge_nested_keys` to `_utils`.
