+++ aiida_optimize/engines/__init__.py
@@ -6,9 +6,10 @@
 This module defines optimization routines to be used with the main optimization WorkChain.
 """
 
+from . import base
 from ._bisection import Bisection
 from ._convergence import Convergence
 from ._nelder_mead import NelderMead
 from ._parameter_sweep import ParameterSweep
 
-__all__ = ['Bisection', 'NelderMead', 'ParameterSweep', 'Convergence']
+__all__ = ['base', 'Bisection', 'NelderMead', 'ParameterSweep', 'Convergence']
