+++ aiida_optimize/helpers.py
@@ -14,14 +14,16 @@ def get_nested_result(output: ty.Dict[str, orm.Node], key: str) -> orm.Node:
     """Helper function to retrieve nested outputs from AiiDA processes.
 
     This function supports the nested key syntax:
+
     - Namespaces are separated by a period
-    - A colon `:` indicates accessing inside an AiiDA `Dict`
+    - A colon ``:`` indicates accessing inside an AiiDA ``Dict``
     - Nested access inside the `Dict` is again separated by a period
 
     Examples:
-    - `'x.y'`: retrieve output `y` in the `x` namespace
-    - `'x.y:a.b`: `x.y` is a `Dict`, and we retrieve its content
-      at `['a']['b']`.
+
+    - ``'x.y'``: retrieve output ``y`` in the ``x`` namespace
+    - ``'x.y:a.b'``: ``x.y`` is a ``Dict``, and we retrieve its content
+      at ``['a']['b']``.
 
     Parameters
     ----------
