+++ doc/source/conf.py
@@ -157,4 +157,7 @@ texinfo_documents = [
 ]
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/': None}
+intersphinx_mapping = {
+    'https://docs.python.org/': None,
+    'https://aiida.readthedocs.io/en/latest/': None
+}
